fix(http): 修正类型转换以确保安全和兼容性
- 修改String.hpp中index函数返回值的类型转换
- 调整Query.cpp中decode函数对字符的静态类型转换
- 更新Request.cpp中setMethod函数内部toupper的参数类型处理
- 修正URL.cpp中对端口号的stoi结果进行uint16_t类型转换
- 增强代码的类型安全和可移植性，避免隐式转换问题

+++ include/darabonba/String.hpp
@@ -90,7 +90,7 @@ public:
 
   static int index(const std::string &s, const std::string &substr) {
     auto ret = s.find(substr);
-    return ret == std::string::npos ? -1 : ret;
+    return ret == std::string::npos ? -1 : static_cast<int>(ret);
   }
 
   static std::string toLower(const std::string &s) {
